requirement.txt
openai
PyPDF2
python-dotenv
selenium
webdriver_manager
